# Conflicts:
#	src/main/java/dan200/computercraft/ComputerCraft.java
#	src/test/java/dan200/computercraft/core/ComputerTestDelegate.java

+++ src/main/resources/data/computercraft/lua/rom/programs/delete.lua
@@ -9,9 +9,18 @@ for i = 1, args.n do
     local files = fs.find(shell.resolve(args[i]))
     if #files > 0 then
         for _, file in ipairs(files) do
-            local ok, err = pcall(fs.delete, file)
-            if not ok then
-                printError((err:gsub("^pcall: ", "")))
+            if fs.isReadOnly(file) then
+                printError("Cannot delete read-only file /" .. file)
+            elseif fs.isDriveRoot(file) then
+                printError("Cannot delete mount /" .. file)
+                if fs.isDir(file) then
+                    print("To delete its contents run rm /" .. fs.combine(file, "*"))
+                end
+            else
+                local ok, err = pcall(fs.delete, file)
+                if not ok then
+                    printError((err:gsub("^pcall: ", "")))
+                end
             end
         end
     else
